Move BearerTokenProvider to auth directory
Moved bearerTokenProvider.ts to auth/bearerTokenProvider.ts and updated all imports to reflect the new location. This improves code organization by grouping authentication-related logic under the auth directory.
