+++ modules/zuora/src/zuoraClient.ts
@@ -1,7 +1,7 @@
 import type { Stage } from '@modules/stage';
 import type { z } from 'zod';
 import { Logger } from '@modules/zuora/logger';
-import { BearerTokenProvider } from './bearerTokenProvider';
+import { BearerTokenProvider } from './auth/bearerTokenProvider';
 import { zuoraServerUrl } from './common';
 import { getOAuthClientCredentials } from './oAuthCredentials';
 import { generateZuoraError } from './zuoraErrorHandler';
