+++ modules/zuora/test/zuoraClientIntegration.test.ts
@@ -5,7 +5,7 @@
  */
 
 import { Logger } from '@modules/zuora/logger';
-import { BearerTokenProvider } from '../src/bearerTokenProvider';
+import { BearerTokenProvider } from '../src/auth/bearerTokenProvider';
 import { getOAuthClientCredentials } from '../src/oAuthCredentials';
 import { ZuoraClient } from '../src/zuoraClient';
 import type { ZuoraSubscription } from '../src/zuoraSchemas';
