+++ modules/zuora/test/zuoraIntegration.test.ts
@@ -4,7 +4,7 @@
  * @group integration
  */
 
-import { BearerTokenProvider } from '../src/bearerTokenProvider';
+import { BearerTokenProvider } from '../src/auth/bearerTokenProvider';
 import { getOAuthClientCredentials } from '../src/oAuthCredentials';
 import { getSubscription } from '../src/subscription';
 import { ZuoraClient } from '../src/zuoraClient';
